Move common stuff to a base class.
